+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Unreleased
 - Support mmult import from reportlab < 3.5.61 (#316).
 - ``<symbol>`` nodes now only appears in documents when they are
   referenced, not on their initial appearance.
+- If only rx or ry is defined on a ``<rect>``, the value is used for both.
 
 1.2.1 (2022-01-27)
 ------------------